Changes for 1.1.0
* Support deleting queries
* Display query list errors
* Update Hubdb
* Add changelog

+++ src/components/query_list.js
@@ -6,20 +6,26 @@ var QueryStore = require('../stores/query_store.js');
 var QueryList = React.createClass({
     mixins: [QueryStore.listenTo, Navigation],
     getInitialState() {
-        return { queries: [] };
+        return { queries: QueryStore.all(), error: QueryStore.error() };
     },
     _onChange() {
-        this.setState({ queries: QueryStore.all() });
+        this.setState({
+          queries: QueryStore.all(),
+          error: QueryStore.error()
+        });
     },
     chooseQuery(q) {
         this.props.chooseQuery(q);
     },
     deleteQuery(q) {
+      if (confirm('Are you sure you want to delete this query?')) {
         actions.deleteQuery(q.path);
+      }
     },
     render() {
-        return <div className='fill-navy-dark dark pad2'>
-          {this.state.queries.map((q, i) =>
+        if (this.state.queries.length && !this.state.error) {
+          return <div className='fill-navy-dark dark pad2'>
+            {this.state.queries.map((q, i) =>
               <div
                 key={q.data.query + '-' + i}
                 className='col12 clearfix contain pad0y'>
@@ -30,12 +36,24 @@ var QueryList = React.createClass({
                 </div>
                 <div className='col9 contain small'>
                   <pre className='truncate fill-dark unround'>{q.data.query}</pre>
-                  {false ? <a
-                    onClick={this.deleteQuery.bind(this, q)}
-                    className='icon trash pin-topright'></a> : ''}
+                  <a onClick={this.deleteQuery.bind(this, q)}
+                    className='icon trash pin-topright pad1'></a>
                 </div>
               </div>)}
-        </div>;
+            </div>;
+        } else if (!this.state.queries.length && this.state.error) {
+          return <div className='fill-navy-dark dark pad2'>
+            <div className='icon close strong center pad1'>
+              Error retrieving queries: {this.state.error}
+            </div>
+          </div>;
+        } else if (this.state.queries.length === 0) {
+          return <div className='fill-navy-dark dark pad2'>
+            <div className='icon info strong center pad1'>
+              No queries saved
+            </div>
+          </div>;
+        }
     }
 });
